Broke apart dataset menu into reusable components
